+++ src/MUSIC.C
@@ -12,14 +12,14 @@ extern char *song;
 extern char music_current;
 extern struct sup setup;
 //=========================================================================
-int music_init(void){
+int16_t music_init(void){
 
     if(!setup.music) return 0;
     music_current=-1;
     return 1;
 }
 //=========================================================================
-void music_play(int num,int override){
+void music_play(int16_t num,int16_t override){
     return; // TODO
     /*
     if(!setup.music) return;
@@ -47,7 +47,7 @@ void music_resume(void){
     */
 }
 //=========================================================================
-int music_is_on(void){
+int16_t music_is_on(void){
     return 0; // TODO
     /*
     if(!setup.music) return 0;
